+++ NEWS.md
@@ -1,3 +1,5 @@
+* Función para bajar y leer bases spss desde Alchemer.
+
 # desuctools 0.1.5.9006
 
 * Aumentar la cantidad de encuestas que descarga la función `sg_get`.