+++ NEWS.md
@@ -5,6 +5,7 @@
 # desuctools 0.1.5.9007
 
 * Se agrega la función `rec_labelled` para facilitar la recodificación de variables en formato labelled.
+* Función para bajar y leer bases spss desde Alchemer.
 
 # desuctools 0.1.5.9006